requirements.txt
beautifulsoup4==4.11.1
fastapi==0.85.1
numpy==1.23.5
openfoodfacts==0.1.3
paho_mqtt==1.6.1
pydantic==1.10.4
requests==2.28.1
SQLAlchemy==1.4.44
uvicorn==0.18.3
